Fix tab switching
When switching between tabs on the previous commit, new cooldowns would only be loaded if they came in their role's list after the number that had already been loaded. This turned out to be because of using the iteration variable for the key in `return <Grid.Column key={...}`.

This commit also removes some unnecessary logging, removes the use of lodash, and renames some things to be more accurately named.